requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
websockets==12.0
redis==5.0.1
qdrant-client==1.11.3
python-dotenv==1.0.0
gunicorn==21.2.0
PyPDF2==3.0.1
requests==2.31.0
psycopg2-binary==2.9.10